pyproject.toml
[project]
name = "flatcad"
version = "0.1.0"
description = "Experimental comparison framework for surface reconstruction methods"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "absl-py>=2.3.1",
    "numpy>=2.2.6",
    "omegaconf>=2.3.0",
    "open3d>=0.19.0",
    "plyfile>=1.1.3",
    "pymcubes>=0.1.6",
    "pyyaml>=6.0.3",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.2",
    "tensorboardx>=2.6.4",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "trimesh>=4.10.1",
]
